Initial working program
The program has a defined lifecycle. It is started by sending `#main` to `SWQuoteOfTheDay class`. The
class holds on to a running instance of the program until the `close` event from the user interface is sent.
At this point the class ends the lifecycle by deallocating the instance.

It is still possible to create more than one instance with `#new` or `#main`.